Planning for a more robust LSP experience
There are some useful features in nvim-lspconfig that I don't have
because I'm using the built-in lsp system. It would be nice to have a
way to list attached enabled LSPs that isn't just `:checkhealth`.

It would also be nice to force-restart LSPs. I've noticed that some
people writing go have issues with gopls not picking up changes to
packages after refactoring and they need to manually restart things. I
haven't ran into that yet but I'll be annoyed when I do.